refactor(controls): a Core/Controls home for shared widgets + CvExpander
Move CvLoadingOverlay from Core/Panes to a new Core/Controls (widgets reusable
across features, not pane-windows), and add CvExpander — an Expander with a
Fluent-style chevron (a "›" rotating to point down when open), templated in
Themes/Generic.xaml (ThemeInfo added). Statistics/Usage keep their loading
overlay through the moved type; the Context tab uses both.
