+++ src/Corsinvest.VisualStudio.Agents/Core/Controls/CvLoadingOverlay.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="Corsinvest.VisualStudio.Agents.Core.Panes.CvLoadingOverlay"
+<UserControl x:Class="Corsinvest.VisualStudio.Agents.Core.Controls.CvLoadingOverlay"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              HorizontalAlignment="Center" VerticalAlignment="Center">
