+++ src/Corsinvest.VisualStudio.Agents/Corsinvest.VisualStudio.Agents.csproj
@@ -198,7 +198,8 @@
     <Compile Include="Utils\ClaudePaths.cs" />
     <Compile Include="Utils\ColorUtils.cs" />
     <Compile Include="Utils\PathHelpers.cs" />
-    <Compile Include="Core\Panes\CvLoadingOverlay.xaml.cs">
+    <Compile Include="Core\Controls\CvExpander.cs" />
+    <Compile Include="Core\Controls\CvLoadingOverlay.xaml.cs">
       <DependentUpon>CvLoadingOverlay.xaml</DependentUpon>
     </Compile>
     <Compile Include="Core\Panes\IPaneControl.cs" />
@@ -363,7 +364,10 @@
     <Page Include="Core\Context\ContextUsageControl.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Core\Panes\CvLoadingOverlay.xaml">
+    <Page Include="Themes\Generic.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Core\Controls\CvLoadingOverlay.xaml">
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Core\Panes\DevInfoDialog.xaml">
